# public/api.yml
openapi: 3.0.0
info:
  title: 'Teachme API'
  version: '0.1'
paths:
  '/line/token/{code}':
    get:
      tags:
        - line
      summary: 'get auth info of line by code'
      description: 'get auth info of line by code'
      operationId: 'App\Http\Controllers\Api\LineController::token'
      parameters:
        -
          name: code
          in: path
          description: 'the code'
          required: true
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineAuthInfo'
  /line/bindings:
    put:
      tags:
        - line
      summary: 'bind account to line account'
      description: 'bind account to line account'
      operationId: 'App\Http\Controllers\Api\LineController::bind'
      requestBody:
        request: 'binding info'
        description: product_request
        required: true
        content:
          application/json:
            schema:
              properties:
                id_token:
                  type: string
              type: object
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - { properties: { data: { $ref: '#/components/schemas/UserInterface' } }, type: object }
    post:
      tags:
        - line
      summary: 'bind account to line account'
      description: 'bind account to line account'
      operationId: 'App\Http\Controllers\Api\LineController::getBindings'
      requestBody:
        request: 'binding info'
        description: product_request
        required: true
        content:
          application/json:
            schema:
              properties:
                id_token:
                  type: string
              type: object
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineAuthInfo'
components:
  schemas:
    AuthInfo:
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expire_in:
          type: integer
        token_type:
          type: string
      type: object
    ApiResponse:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
    LineAuthInfo:
      allOf:
        -
          $ref: '#/components/schemas/AuthInfo'
        -
          properties:
            id_token:
              type: string
    UserInterface:
      allOf:
        -
          properties:
            id:
              type: string
            name:
              type: string
            school_id:
              description: 'null if is principal'
              type: integer
            username:
              description: 'not exists if is teacher'
              type: string
            line_user_id:
              type: string
          type: object